third_party/WebKit/LayoutTests/hittesting/border-hittest.html

<!doctype html>
<html>
<head>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<style>
img{
    background-color:red;
    border-radius: 50px;
}
</style>
</head>
<body>
<img src="../fast/borders/resources/mask.png" width="100px" height="100px" />
</body>
<script>
test(function(t)
{
    var element = document.elementFromPoint(100, 100);
    assert_equals(element.nodeName, 'BODY');
}, "elementFromPoint should return BODY when outside an element");
</script>
</html>
